pynapple/requirements.txt
flask~=3.0
gunicorn~=21.0
flask~=3.0
flask-sqlalchemy~=3.1
psycopg2-binary~=2.9
pymysql~=1.1
cryptography~=42.0  # needed by mysql but for some reason not implicitly installed
Flask-Migrate~=4.0
redis~=5.0
python-memcached~=1.62
